requirements.txt
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
pycodestyle==2.6.0
python-dotenv==0.14.0
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
